requirements.txt
requests
python-dotenv
pandas
influxdb-client
influxdb
yfinance